+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## v1.0.7
+
+### New Features:
+- Custom Emoji Particles**: Added support for emoji particles alongside image particles. Users can now use any emojis as falling particles, creating more dynamic visual effects.
+- Enhanced performance and responsiveness for dynamic updates of particle properties, ensuring smoother interactions.
+
 ## v1.0.6
 
 ### New Features: